Tweak the export_prefix behavior, and add some messaging
* export_prefix didn't seem to be working without a builder, for me, but
adding that simple one did the trick.
* 'schema','rs','resultset',and 'rset' are not prefixed in the older
plugin. Not prefixing them here makes this a drop-in replacement
(almost; see caveats below) for Dancer2::Plugin::DBIC. I turned off
prefixing for those words, but it seems to me that should be switchable,
so that older apps can use it that way, but newer apps (or those using
this plugin multiple times) can prefix.  I'll take care of a
"legacy_noprefix" config parameter later.
* No need to add 'resultset' a second time; it's already being dealt
with.
* Add some croakery if the user fails to properly configure the plugin,
rather than giving a bunch of failures in Class::C3::Componentised

CAVEATS:

If you have a table named the same thing as a dsl keyword in Dancer2--
the one I found was "session", then you've got to either adjust
Dancer2::Session's parameters, or use prefixing.  That
particular conflict causes some pretty mad chaos.  (note for the docs)

This plugin cause Dancer2::Session::DBIC to break if you've used
"provider: DBIC" without giving it explicit connection details--which is
normal behavior.  We need to deal with it so that if someone is using
this plugin, it'll properly pass the schema to Dancer2::Session without
writing a whole new provider.

Ditto for Dancer2::Plugin::Auth::Extensible::Provider::DBIC.  Both seem
to be searching for D2::P::DBIC, and if it finds them, pass the schema
along, no worries.  It shouldn't be hard to either a) make this one
masquerade, or b) tell them it's an either-or situation with small PRs